reworked clipping to parent

+++ src/gui/position_info.h
@@ -5,7 +5,7 @@ namespace gui
 {
     struct Position_info
     {
-        sf::Vector2i offset;
+        sf::Vector2f offset;
         sf::Vector2f percentage_offset;
         sf::Vector2f relative_to;
     };
